requirements_dev.txt
bumpversion==0.5.3
wheel==0.30.0
djangorestframework==3.11.0
